add interface

+++ src/main/java/org/qxdn/birthdayreminder/repository/CharacterRepository.java
@@ -5,10 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.Date;
 import java.util.List;
 
-public interface CharacterRepository extends JpaRepository<CharacterDO, Integer> {
+public interface CharacterRepository extends JpaRepository<CharacterDO, Long> {
 
     @Query("SELECT c FROM CharacterDO c WHERE c.birthMonth = :month AND c.birthDay = :day")
     List<CharacterDO> findCharactersWithBirthday(@Param("month") Integer month, @Param("day") Integer birthday);
